// tree-world/api/model/api-add-task-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { ApiTaskInfo } from './api-task-info';

/**
 * 
 * @export
 * @interface ApiAddTaskResponse
 */
export interface ApiAddTaskResponse {
    /**
     * 
     * @type {ApiTaskInfo}
     * @memberof ApiAddTaskResponse
     */
    'task'?: ApiTaskInfo;
}